refactor: part 1 of 2 re: #102; creates & populates mdi_source
